update more tests
